Extract common Makefile parts into Makefile.COMMON.
This is a test-drive for separating all common build functionality into
a separate Makefile.COMMON, which is included from an individual
project's Makefile (mainly for the various exporters, which should have
almost identical Makefiles, modulo accrued inconsistencies).

The main Makefile then only needs to define the VERSION and TARGET
variables and the Makefile.COMMON defines all the common targets. While
doing this, I decided per variable which one should be simply expanded
(`:=`) vs. conditionally set (`?=`). The former allows evaluating values
only once and on the spot (good for potentially costly evaluations),
while the latter allows overriding variables from the parent Makefile.
Unfortunately it's not possible to combine the two variants (like
`?:=`), so I had to chose which one seemed more important for each var
(being able to override it or only evaluating it once).

There's the question about how the Makefile.COMMON should be maintained
and how it should make its way into the various *_exporter (etc.) repos
in the future, and how it should be kept up-to-date. The authoritative
version of this file should probably live in its own ("build_common"?)
repo. It would of course be nice to download it dynamically from a
project's specific Makefile itself, but that would mean that all the
targets that Makefile.COMMON defines would only become available once
it's actually downloaded, so this would always require two make steps,
which doesn't seem like a good option. Another option could be to simply
manually copy it into every repository when there are upstream changes
to the common file. Opinions/ideas around this are appreciated.

One thing I generally improved in this new Makefile is that targets
like $(BINARY) are really only built if needed - that is, if one of the
source files actually changed. This is solved by creating a
"dependencies-stamp" file so we don't always repeat downloading
dependencies and "go build" every time you call "make", even if
the code didn't change. Granted, remote dependencies could have changed
upstream between make runs, but for those to take effect, I think it's
fair to require a "make clean" or deleting the "dependencies-stamp"
file.

I also simplified and automated the code around choosing the correct Mac
OS X Go release version, but wasn't able to test it. If anyone with OS X
could test it, that would be great.

Also, PHONY targets are now declared as such inline, making it more
obvious which target is PHONY, and easier to maintain PHONY-ness in
general.