Fix timeout and keepalive by using Client.Timeout
With net.Conn.SetDeadline the connection could not be reused for
additional requests after the timeout, and even a fast response could
time out while reading, because the timer was started by a previous
request.

With a five second scrape interval, the log will fill up with this
error:

   ERRO[0108] Unexpected error while reading CSV: read tcp n.n.n.n:57648->m.m.m.m:443: i/o timeout source=haproxy_exporter.go:331

With this change I can not see any more errors and by using netstat I
can see that the same port/connection is reused for a while.