support conn_tot stats
haproxy 1.7 introduces a distinction between connection and sessions,
sessions represents an end to end connection:
"with the introduction of SSL, proxy protocol and layer4 ACLs it was
needed to cut the end-to-end sessions in smaller parts, hence the
introduction of 'connections'"

this patch renames the old `connections_total` to `sessions_total` in order
to use the same vocabulary as haproxy; I was puzzled to keep the compatibility
but in the long term I think this will introduces more incomprehension
than it resolves.
so this is a breaking change.

`connections_total` now represents the new metric from haproxy.

also add comment to the new available metrics.

Signed-off-by: William Dauchy <w.dauchy@criteo.com>